Change XmlFlowAnnotationSource implementation to expose CustomAttributes (#1216)
- Change XmlFlowAnnotationSource to expose CustomAttributes
- Use ICustomAttributeProvider to simplify when asking for information
about annotations
- Assembly name now could be '*' in the assembly element, there is a new assembly property inside the attribute element, leaving no assembly name in this property means to look for the type in all the loaded assemblies.
- Added new error codes, swapped non used codes for new error codes
- Changed documentation to reflect the new functionalities
- Move attribute sources to the main linker namespace
