+++ test/Mono.Linker.Tests.Cases/DataFlow/XmlAnnotations.xml
@@ -3,54 +3,54 @@
   <assembly fullname="Test">
     <type fullname="Mono.Linker.Tests.Cases.DataFlow.XmlAnnotations">
       <field name="_typeWithDefaultConstructor">
-        <attribute fullname="System.DoesNotExistattribute">
+        <attribute fullname="System.DoesNotExistattribute" assembly="Mono.Linker.Tests.Cases.Expectations">
           <argument>0</argument>
         </attribute>
-        <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers">
+        <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembersAttribute">
           <argument>DefaultConstructor</argument>
         </attribute>
       </field>
       <method signature="System.Void TwoAnnotatedParameters(System.Type,System.Type)">
         <parameter name="type">
-          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers">
+          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembersAttribute" assembly="Mono.Linker.Tests.Cases.Expectations">
             <argument>DefaultConstructor</argument>
           </attribute>
         </parameter>
         <parameter name="type2">
-          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers">
+          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembersAttribute" assembly="Mono.Linker.Tests.Cases.Expectations">
             <argument>PublicConstructors</argument>
           </attribute>
         </parameter>
       </method>
       <method signature="SpacesBetweenParametersWrongArgument (System.Type, System.Boolean)">
         <parameter name="type">
-          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers">
+          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembersAttribute" assembly="Mono.Linker.Tests.Cases.Expectations">
             <argument>NonValidArgument</argument>
           </attribute>
         </parameter>
       </method>
       <method signature="GenericMethod&lt;T&gt;(T,System.Type)">
         <parameter name="type">
-          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers">
+          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembersAttribute" assembly="Mono.Linker.Tests.Cases.Expectations">
             <argument>DefaultConstructor</argument>
           </attribute>
         </parameter>
       </method>
       <method name="ReturnConstructorsFailure">
         <return>
-          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers">
+          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembersAttribute" assembly="Mono.Linker.Tests.Cases.Expectations">
             <argument>DefaultConstructor</argument>
           </attribute>
         </return>
       </method>
       <property name="PropertyWithDefaultConstructor">
-        <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers">
+        <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembersAttribute" assembly="Mono.Linker.Tests.Cases.Expectations">
           <argument>DefaultConstructor</argument>
         </attribute>
       </property>
       <type name="NestedType">
         <field name="_typeWithDefaultConstructor">
-          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers">
+          <attribute fullname="System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembersAttribute" assembly="Mono.Linker.Tests.Cases.Expectations">
             <argument>DefaultConstructor</argument>
           </attribute>
         </field>
